% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDIRWithNoReplicate.R
\name{.generateInteractionMatrixCount}
\alias{.generateInteractionMatrixCount}
\title{.generateInteractionMatrixCount}
\usage{
.generateInteractionMatrixCount(file, sample)
}
\arguments{
\item{file}{a character specifying the input file
The inputfile is a interaction matrix}

\item{sample}{a character specifying the sample type, i.e. treat or control}
}
\value{
a dataframe containing the interaction index and count
}
\description{
Get the interaction count dataframe
}
