% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDIRWithNoReplicate.R
\name{GetDIRWithNoReplicate}
\alias{GetDIRWithNoReplicate}
\title{GetDIRWithNoReplicate}
\usage{
GetDIRWithNoReplicate(chr, treat, ctrl, bcv, output)
}
\arguments{
\item{chr}{a character string specifying the interaction regions from which chromosome}

\item{treat}{a character string specifying the name and path of the treatment file (.csv)}

\item{ctrl}{a character string specifying the name and path of the control file (.csv)}

\item{bcv}{a number specifying the hypothetical biological covaraiance, default is 0.4
See section 2.12
https://bioconductor.org/packages/release/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf}

\item{output}{A character string specifying the name and path of output file
the default path is as same as the treatment file
the default output file name is the treatment file name prefixed with "DIR_"}
}
\value{
a dataframe of the output data
}
\description{
This function reads interaction counts and performs differential analysis between two conditions.
This function is specifically for experiments without biological replicate.
Hence the result has no statistical significance.
Please treat the log2 fold change, p value, or FDR as descriptive values.
}
\examples{
\dontrun{
df <- GetDIRWithNoReplicate(
chr = "chr2",
treat = here::here("./Results/processing/44112_A_bg_43615_mc6contact_map_extracted.csv"),
ctrl = here::here("./Results/processing/44111_A_ctrl_43614_mc5contact_map_extracted.csv"),
bcv = 0.4,
output = NULL
)
}
}
\author{
Tsunghan Hsieh
}
