% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractionPosIndex.R
\name{getInteractionPosIndex}
\alias{getInteractionPosIndex}
\title{getInteractionPosIndex}
\usage{
getInteractionPosIndex(chr, start, end, res)
}
\arguments{
\item{chr}{a character string specifying the interested segment from which chromosome}

\item{start}{a integer string specifying starting position of the interested segment}

\item{end}{a integer string specifying starting position of the interested segment}

\item{res}{a integer specifying the resolution
The above parameters should be same as ./Python/subset_hic_data.py}

\item{output}{A character string specifying the name and path of output file}
}
\value{
a dataframe of the output data
}
\description{
This function generates the position index for the interested genomic segment and write into a .txt file.
The position index is the mid-point of the genomic regions
For example, if there is a interaction between two genomic regions: 11000-16000 and 18000-23000
Then the index will be 13500 - 20500
}
\examples{
\dontrun{
df <- getInteractionPosIndex(
chr = "chr2",
start = 112735986,
end = 113204585,
res = 5000,
output = here::here("./Results/processing/start_position_index.txt")
)
}
}
\author{
Tsunghan Hsieh
}
