% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDIRWithNoReplicate.R
\name{.generateInteractionMatrixCount}
\alias{.generateInteractionMatrixCount}
\title{.generateInteractionMatrixCount}
\usage{
.generateInteractionMatrixCount(file, start_position_index)
}
\arguments{
\item{file}{a character specifying the input file
The inputfile is a interaction matrix}

\item{start_position_index}{a dataframe containing the starting position of the each interaction}
}
\value{
a dataframe containing the interaction index and count
}
\description{
Get the interaction count dataframe
}
